+++ src/ds_caselaw_utils/courts.py
@@ -285,31 +285,30 @@ class CourtsRepository:
                 groups.append(CourtGroup(category.get("display_name"), courts))
         return groups
 
-    def get_grouped_selectable_courts(self) -> list[CourtGroup]:
+    def get_grouped_institutions_with_property(
+        self, institution_type: InstitutionType, court_property: str
+    ) -> list[CourtGroup]:
         groups = []
         for category in self._data:
-            if not category.get("is_tribunal"):
+            if category.get("is_tribunal", False) == (institution_type == InstitutionType.TRIBUNAL):
                 courts = [
-                    Court(court, type=InstitutionType.COURT)
-                    for court in category.get("courts", [])
-                    if court["selectable"]
+                    Court(court, institution_type) for court in category.get("courts", []) if court.get(court_property)
                 ]
                 if len(courts) > 0:
                     groups.append(CourtGroup(category.get("display_name"), courts))
         return groups
 
+    def get_grouped_selectable_courts(self) -> list[CourtGroup]:
+        return self.get_grouped_institutions_with_property(InstitutionType.COURT, "selectable")
+
     def get_grouped_selectable_tribunals(self) -> list[CourtGroup]:
-        groups = []
-        for category in self._data:
-            if category.get("is_tribunal"):
-                courts = [
-                    Court(court, InstitutionType.TRIBUNAL)
-                    for court in category.get("courts", [])
-                    if court["selectable"]
-                ]
-                if len(courts) > 0:
-                    groups.append(CourtGroup(category.get("display_name"), courts))
-        return groups
+        return self.get_grouped_institutions_with_property(InstitutionType.TRIBUNAL, "selectable")
+
+    def get_grouped_listable_courts(self) -> list[CourtGroup]:
+        return self.get_grouped_institutions_with_property(InstitutionType.COURT, "listable")
+
+    def get_grouped_listable_tribunals(self) -> list[CourtGroup]:
+        return self.get_grouped_institutions_with_property(InstitutionType.TRIBUNAL, "listable")
 
     def get_listable_groups(self) -> list[CourtGroup]:
         groups = []